issue-49 Move sorted array to separate file + add factory method
